refactor: Login and register now share styles (reduce repetition)
